requirements.txt
fastapi==0.78.0
uvicorn[standard]==0.18.2
sqlalchemy==1.4.35
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-jose[cryptography]==3.3.0
requests==2.28.1
pydantic[email]==1.9.0
python-multipart==0.0.6
alembic==1.7.7
python-keycloak==3.7.0